test(integration): consolidate duplicate _insert_event helpers and unique-ify camera names
Two _insert_event definitions existed in test_api.py (lines 190 and 467).
Python's sequential rebinding meant the second (zones-aware) silently
shadowed the first for all callers; the first was dead code. The two
signatures also differed (end_ts had a default in the first, not the
second), making the shadowing non-obvious.

Fix:
- Move _insert_segment and _insert_event to conftest.py as plain module-
  level helpers. Unified signature includes all kwargs from both versions:
  _insert_event(db_path, camera, start_ts, end_ts=None, label, zones).
- Remove both definitions from test_api.py; import from conftest instead.
- Rename generic camera names to test-function-scoped names ('test-cam'
  → 'cameras-test-cam', 'evt-range-cam' → 'timeline-evt-range-cam') so
  rows cannot bleed across tests if fixture scope is ever elevated.
- Add comment on test_app fixture documenting the scope and camera-name
  uniqueness requirement (F-15).

No behavior change. 38 integration tests pass.
